Wheeee

+++ service/icharger_controller.cpp
@@ -44,7 +44,7 @@ struct ChannelStatusJson {
             
             QVariantMap cell;
             cell[STR_CHANNEL_STATUS_CELL_NUMBER] = index;
-            cell[STR_CHANNEL_STATUS_CELL_VOLTAGE] = c.cell_voltage[index] / 1000.0;
+            cell[STR_CHANNEL_STATUS_CELL_VOLTAGE] = ((double)c.cell_voltage[index]) / 1000.0;
             cell[STR_CHANNEL_STATUS_CELL_BALANCE_STATUS] = c.balance_status[index];
             cell[STR_CHANNEL_STATUS_CELL_RESISTANCE] = c.cell_resistance[index];
             cells << cell;
